requirements.txt

grpcio
grpcio-tools
jsonify